Creating Subfolders for LIS with SegmentTree
